authentification avec  JWT etape 1  add  line : "require": {           "php": ">=5.5.9",           "laravel/framework": "5.1.*",      +  "tymon/jwt-auth": "0.5.*"    },    etape 2 :    cmd>composer update    etape 3 :    in config/app.php add line    Tymon\JWTAuth\Providers\JWTAuthServiceProvider::class    etape 4 :  in config/app.php    add line in array aliases     'JWTAuth' => Tymon\JWTAuth\Facades\JWTAuth::class,            'JWTFactory' => Tymon\JWTAuth\Facades\JWTFactory::class,  etape 5 :  cmd  php artisan vendor:publish --provider="Tymon\JWTAuth\Providers\JWTAuthServiceProvider"  php artisan jwt:generate  Partie AngularJS  cmd            cd public            npm install angular satellizer angular-ui-router bootstrap  etape  2 fichier Kernel  commanter la ligne   // AppHttpMiddlewareVerifyCsrfToken::class,   ajouter la ligne   'jwt.auth' => Tymon\JWTAuth\MiddlewareGetUserFromToken::class,

+++ app/Http/Controllers/MeetingController.php
@@ -2,16 +2,20 @@
 
 namespace App\Http\Controllers;
 
-use App\Meeting;
+//use Carbon\Carbon\Carbon;
 use Illuminate\Http\Request;
 
 use App\Http\Requests;
 
+use App\Meeting;
+use JWTAuth;
 class MeetingController extends Controller
 {
     public function __construct()
     {
-        // $this->middleware('name');
+         $this->middleware('jwt.auth',['only'=>[
+             'update','store','destroy']
+         ]);
     }
 
     /**
@@ -43,30 +47,39 @@ class MeetingController extends Controller
      */
     public function store(Request $request)
     {
-//        $this->validate($request, [
-//            'title' => 'required',
-//            'description' => 'required',
-//            'time' => 'required|date_format:YmdHie'
-//        ]);
-
+        if(!$user = JWTAuth::parseToken()->authenticate()){
+            return response()->json(['msg'=>'User not found'],404);
+        }
         $title = $request->input('title');
         $description = $request->input('description');
         $time = $request->input('time');
-        $user_id = $request->input('user_id');
+        $user_id = $user->id;
 
         $meeting = new Meeting([
             'title' => $title,
             'description' => $description,
-            'time' => $time,
-            'user_id' => $user_id]);
-        $meeting->save();
+            'time' => $time
+        ]);
+        if ($meeting->save()) {
+            $meeting->users()->attach($user_id);
+            $meeting->view_meeting = [
+                'href' => 'api/v1/meeting/' . $meeting->id,
+                'method' => 'GET',
+            ];
+            $message = [
+                'msg' => 'Meeting created',
+                'meeting' => $meeting
+            ];
 
+            return response()->json($message, 201);
+
+        }
         $response = [
-            'msg' => 'Meeting created',
-            'meeting' => $meeting
+            'msg' => 'Error during creation'
         ];
 
-        return response()->json($response, 201);
+        return response()->json($response,404);
+
     }
 
     /**
@@ -104,26 +117,36 @@ class MeetingController extends Controller
      */
     public function update(Request $request, $id)
     {
+        $this->validate($request, [
+            'title' => 'required',
+            'description' => 'required',
+            'time' => 'required|date_format:YmdHie'
+        ]);
+        if(!$user = JWTAuth::parseToken()->authenticate()){
+            return response()->json(['msg'=>'User not found'],404);
+        }
         $title = $request->input('title');
         $description = $request->input('description');
         $time = $request->input('time');
-        $user_id = $request->input('user_id');
-        $meeting = [
-            'title' => $title,
-            'description' => $description,
-            'time' => $time,
-            'user_id' => $user_id,
-            'view_meeting' => [
-                'href' => 'api/v1/meeting/1',
-                'method' => 'GET'
-            ]
+        $user_id = $user->id;
+        $meeting = Meeting::with('users')->findOrFail($id);
+        if (!$meeting->users()->where('users.id', $user_id)->first()) {
+            return response()->json(['msg' => 'user not registered for meeting, update not successful'], 401);
+        }
+        $meeting->time = Carbon::createFromFormat('YmdHie', $time);
+        $meeting->title = $title;
+        $meeting->description = $description;
+        if (!$meeting->update()) {
+            return response()->json(['msg' => 'Error during updating']);
+        }
+        $meeting->view_meeting = [
+            'href' => 'api/v1/meeting/' . $meeting->id,
+            'method' => 'GET'
         ];
-
         $response = [
             'msg' => 'Meeting updated',
             'meeting' => $meeting
         ];
-
         return response()->json($response, 200);
     }
 
@@ -135,19 +158,23 @@ class MeetingController extends Controller
      */
     public function destroy($id)
     {
+
         $meeting = Meeting::findOrFail($id);
-//        if(!$user = JWTAuth::parseToken()->authenticate()){
-//            return response()->json(['msg'=>'User not found'],404);
-//        }
-//        if (!$meeting->users()->where('users.id', $user->id)->first()) {
-//            return response()->json(['msg' => 'user not registered for meeting, update not successful'], 401);
-//        }
-//        $users =  $meeting->users;
-//        $meeting->users()->detach();
+
+        if(!$user = JWTAuth::parseToken()->authenticate()){
+            return response()->json(['msg'=>'User not found'],404);
+        }
+
+        if (!$meeting->users()->where('users.id', $user->id)->first()) {
+            return response()->json(['msg' => 'user not registered for meeting, update not successful'], 401);
+        }
+
+        $users =  $meeting->users;
+        $meeting->users()->detach();
         if(!$meeting->delete()){
-//            foreach ($users as $user) {
-//                $meeting->users()->attach($user);
-//            }
+            foreach ($users as $user) {
+                $meeting->users()->attach($user);
+            }
             return response()->json(['msg'=>'deletion failed'],404);
         }
         $response = [
