+++ app/Http/Controllers/RegistrationController.php
@@ -1,13 +1,16 @@
 <?php
-
 namespace App\Http\Controllers;
-
+use App\Meeting;
+use App\User;
 use Illuminate\Http\Request;
-
 use App\Http\Requests;
-
+use JWTAuth;
 class RegistrationController extends Controller
 {
+    public function __construct()
+    {
+        $this->middleware('jwt.auth');
+    }
     /**
      * Store a newly created resource in storage.
      *
@@ -16,59 +19,56 @@ class RegistrationController extends Controller
      */
     public function store(Request $request)
     {
-
+        $this->validate($request, [
+            'meeting_id' => 'required',
+            'user_id' => 'required',
+        ]);
         $meeting_id = $request->input('meeting_id');
         $user_id = $request->input('user_id');
-
-        $meeting = [
-            'title' => 'Title',
-            'description' => 'Description',
-            'time' => 'Time',
-            'view_meeting' => [
-                'href' => 'api/v1/meeting/1',
-                'method' => 'GET'
+        $meeting = Meeting::findOrFail($meeting_id);
+        $user    = User::findOrFail($user_id);
+        $message = [
+            'msg' => 'User is already registered for meeting',
+            'user' => $user,
+            'meeting' => $meeting,
+            'unregister' => [
+                'href' => 'api/v1/meeting/registration/' . $meeting->id,
+                'method' => 'DELETE'
             ]
         ];
-
-        $user = [
-            'name' => 'Name'
-        ];
-
+        if($meeting->users()->where('users.id',$user->id)->first()){
+          return response()->json($message,404);
+        };
+        $user->meetings()->attach($meeting);
         $response = [
             'msg' => 'User registered for meeting',
             'meeting' => $meeting,
             'user' => $user,
             'unregister' => [
-                'href' => 'api/v1/meeting/registration/1',
+                'href' => 'api/v1/meeting/registration/'.$meeting->id,
                 'method' => 'DELETE'
             ]
         ];
-
         return response()->json($response, 201);
     }
-
     /**
      * Remove the specified resource from storage.
      *
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
+    //http://localhost:8000/api/v1/meeting/registration/4?token=eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJzdWIiOjUsImlzcyI6Imh0dHA6XC9cL2xvY2FsaG9zdDo4MDAwXC9hcGlcL3YxXC91c2VyXC9zaWduaW4iLCJpYXQiOjE0Njk5MTg0MDIsImV4cCI6MTQ2OTkyMjAwMiwibmJmIjoxNDY5OTE4NDAyLCJqdGkiOiI4YmQwZTA0NThjZDYwNTIyM2YyMjNkNzZmNmFkZTEzNCJ9.znQunX5ZAq31kqkzSDegdqic9gFrMKwYC4gLCw4m70Y
+    //http://localhost:8000/api/v1/meeting/registration/2?token=eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJzdWIiOjUsImlzcyI6Imh0dHA6XC9cL2xvY2FsaG9zdDo4MDAwXC9hcGlcL3YxXC91c2VyXC9zaWduaW4iLCJpYXQiOjE0Njk5MTg0MDIsImV4cCI6MTQ2OTkyMjAwMiwibmJmIjoxNDY5OTE4NDAyLCJqdGkiOiI4YmQwZTA0NThjZDYwNTIyM2YyMjNkNzZmNmFkZTEzNCJ9.znQunX5ZAq31kqkzSDegdqic9gFrMKwYC4gLCw4m70Y
     public function destroy($id)
     {
-        $meeting = [
-            'title' => 'Title',
-            'description' => 'Description',
-            'time' => 'Time',
-            'view_meeting' => [
-                'href' => 'api/v1/meeting/1',
-                'method' => 'GET'
-            ]
-        ];
-
-        $user = [
-            'name' => 'Name'
-        ];
-
+        $meeting = Meeting::findOrFail($id);
+        if(!$user = JWTAuth::parseToken()->authenticate()){
+            return response()->json(['msg'=>'User not found'],404);
+        }
+        if (!$meeting->users()->where('users.id', $user->id)->first()) {
+            return response()->json(['msg' => 'user not registered for meeting, delete operation not successful'], 401);
+        }
+        $meeting->users()->detach($user->id);
         $response = [
             'msg' => 'User unregistered for meeting',
             'meeting' => $meeting,
@@ -79,7 +79,6 @@ class RegistrationController extends Controller
                 'params' => 'user_id, meeting_id'
             ]
         ];
-
         return response()->json($response, 200);
     }
 }
